pyproject.toml
[tool.pdm]

[tool.pdm.scripts]
db-makemigration = {cmd = "pdm run alembic -c identity_service/database/alembic.ini revision --autogenerate -m \"Initial migration\""}
db-upgrade = {cmd = "pdm run alembic -c identity_service/database/alembic.ini upgrade heads"}
dev = {cmd = "uvicorn identity_service:app --host 0.0.0.0 --port 8080 --reload"}

[tool.pdm.dev-dependencies]
dev = [
    "mypy>=1.2.0",
    "pylint>=2.17.3",
    "black>=23.3.0",
    "pytest>=7.3.1",
    "ipdb>=0.13.13",
]
[project]
name = "identity-service"
version = "1.0.0"
description = ""
authors = [
    {name = "Marek Leszczynski", email = "marek.m.leszczynski@gmail.com"},
]
dependencies = [
    "fastapi>=0.95.1",
    "sqlalchemy>=2.0.11",
    "pydantic>=1.10.7",
    "argon2-cffi>=21.3.0",
    "aiohttp>=3.8.4",
    "cryptography>=40.0.2",
    "alembic>=1.10.4",
    "psycopg2-binary>=2.9.6",
    "uvicorn>=0.22.0",
    "jwcrypto>=1.4.2",
]
requires-python = ">=3.11"
license = {text = "GNU GPLv3"}

[tool.pylint.SIMILARITIES]
ignore-imports=true

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "missing-function-docstring",
    "missing-class-docstring",
    "missing-module-docstring",
    "too-few-public-methods"
]
ignore = "migrations"
extension-pkg-whitelist = ["pydantic"]

[tool.isort]
profile = "black"
known_first_party = ["database", "common", 'schemas']

[tool.mypy]
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_any_explicit = false
no_implicit_optional = true
ignore_missing_imports = true
exclude = "migrations"
plugins = ["pydantic.mypy"]